Restructure conftest
