+++ tests/schains/config/node_info_test.py
@@ -76,7 +76,7 @@ def test_generate_current_node_info(
             node_id=1,
             ecdsa_key_name='123',
             static_node_info=static_node_info,
-            schain=get_schain_struct(schain_name=_schain_name),
+            schain=get_schain_struct(_test_schain_name=_schain_name),
             rotation_id=0,
             nodes_in_schain=4,
             schain_base_port=10000,
@@ -100,7 +100,7 @@ def test_generate_current_node_info(
             node_id=1,
             ecdsa_key_name='123',
             static_node_info=static_node_info,
-            schain=get_schain_struct(schain_name=_schain_name),
+            schain=get_schain_struct(_test_schain_name=_schain_name),
             rotation_id=0,
             nodes_in_schain=4,
             schain_base_port=10000,
