+++ tests/schains/monitor/containers_test.py
@@ -12,7 +12,7 @@ def test_monitor_skaled_container(
     schain_db, skaled_status, dutils, ssl_folder, schain_config, cleanup_schain_containers
 ):
     schain_record = upsert_schain_record(schain_db)
-    schain = get_schain_struct(schain_name=schain_db)
+    schain = get_schain_struct(_test_schain_name=schain_db)
 
     monitor_skaled_container(schain.name, schain_record, skaled_status, dutils=dutils)
     assert not is_container_exists(schain_db, dutils=dutils)
@@ -31,7 +31,7 @@ def test_monitor_skaled_container_exit_time_reached(
     cleanup_schain_containers,
 ):
     schain_record = upsert_schain_record(schain_db)
-    schain = get_schain_struct(schain_name=schain_db)
+    schain = get_schain_struct(_test_schain_name=schain_db)
 
     with mock.patch('core.monitor.schain.containers.is_volume_exists', return_value=True):
         schain_record.set_failed_rpc_count(100)
@@ -59,7 +59,7 @@ def test_monitor_skaled_container_ec(
     schain_db, skaled_status, dutils, ssl_folder, schain_config, cleanup_schain_containers
 ):
     schain_record = upsert_schain_record(schain_db)
-    schain = get_schain_struct(schain_name=schain_db)
+    schain = get_schain_struct(_test_schain_name=schain_db)
 
     run_custom_schain_container(dutils, schain.name, entrypoint=['sh', 'exit', '1'])
     # To make sure container initializaed
